% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccupationsNZ1891.R
\docType{package}
\name{OccupationsNZ1891}
\alias{OccupationsNZ1891}
\alias{OccupationsNZ1891-package}
\title{New Occupations in 1891}
\description{
Occupations as reported in the 1891 New Zealand census broken down by gender and age group
}
\details{
OccupationsNZ1891
A data package with occupation data from the 1891 census.

Use \code{data(package='OccupationsNZ1891')$results[, 3]} tosee a list of availabledata sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{occupations}
}
