% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccupationsNZ1891.R
\docType{data}
\name{occupations}
\alias{occupations}
\title{1891 Occupation data}
\format{a \code{tbl_df} containing the following fields:
\describe{
\item{Class}{}
\item{Sub Class}{}
\item{Order}{}
\item{Sub Order}{}
\item{Occupations}{}
\item{Gender}{}
\item{Age Group}{}
\item{Count}{}
}}
\source{
The data comes from https://www3.stats.govt.nz/historic_publications//1891-census/1891-results-census/1891-results-census.html?_ga=2.3235056.1979237612.1536527586-1395609890.1499578343
}
\description{
Detailed description of the data
}
\seealso{
\link{OccupationsNZ1891}
}
